+++ src/modules/product/product.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
 import { Model } from 'mongoose';
 import { validateOrReject } from 'class-validator';
@@ -6,12 +6,14 @@ import { validateOrReject } from 'class-validator';
 import { Product, UpdateProduct } from './product.interface';
 import { CreateProductDto } from './dto/create-product.dto';
 import { UpdateProductDto } from './dto/update-product.dto';
+import { InvoiceItemService } from '../invoice/invoice-item/invoice-item.service';
 
 
 @Injectable()
 export class ProductService {
   constructor(
     @InjectModel('Product') private readonly productModel: Model<Product>,
+    private readonly invoiceItemService: InvoiceItemService,
   ) {
   }
 
@@ -26,14 +28,12 @@ export class ProductService {
   async create(product: CreateProductDto): Promise<Product> {
     const productDto = new CreateProductDto(product);
     await validateOrReject(productDto);
-
     return this.productModel.create(productDto);
   }
 
   async update(query: UpdateProduct, product: UpdateProductDto): Promise<Product> {
     const productDto = new UpdateProductDto(product);
     await validateOrReject(productDto);
-
     return this.productModel.findOneAndUpdate(
       query,
       product,
@@ -42,6 +42,10 @@ export class ProductService {
   }
 
   async delete(id: string): Promise<Product> {
+    const invoiceItems = await this.invoiceItemService.getByProductId(id);
+    if (invoiceItems.length) {
+      throw new HttpException('Product exsists in invoices', HttpStatus.BAD_REQUEST);
+    }
     return this.productModel.findByIdAndDelete(id);
   }
 }
